deploy/setup/requirements.txt
minio>=7.2.3
python-dotenv>=1.0.1
ldap3
requests
docker